common-lib: fix data type of SCU bus slave

+++ modules/common-libs/include/common-fwlib.h
@@ -108,7 +108,7 @@ volatile uint32_t * fwlib_getMilPiggy();
 volatile uint32_t * fwlib_getOLED();
 
 // get WB address of SCU bus master
-volatile uint16_t * fwlib_getSbMaster();
+volatile uint32_t * fwlib_getSbMaster();
 
 // acquire and publish NIC data
 void fwlib_publishNICData();
